Merged PR 6: Add validation error return from API by HTTP 400
Adds validation of currently-implemented commands via FluentValidation and a MediatR pipeline Validation behavior filter.

Request validation failures are collected and put into a custom ValidationException type which is thrown.

An exception filter attribute has been created and added to the Asp.net core action invocation pipeline, which inspects and handles certain types of exceptions. In this case it catches the ValidationException and returns to the API caller an Http 400 action result with the list of validation errors included.

The next step will be for the end-user client to handle the HTTP 400 and present the list of validation errors to the user.

As part of the implementation the following additional refactors were performed:
- Organized existing Members commands and queries to a new sub-folder tree in the project.
- Moved dependency injection for referenced projects from the Startup.cs file to a file called DependencyInjection.cs in each respective project. This simplifies the Startup file and extracts unnecessary references from the startup (UI) project.

Related work items: #87