Merged PR 15: Add modal with server-side member validation errors
Add a modal service and then wraps the API call in the Syncfusion custom data adaptor with code that catches http response errors and, when appropriate, presents a modal dialog containing a list of the field validation issues to the user.

Known issue: When inserting a new row and an http 400 is returned, the row currently being edited is lost and the user, without prompting, must click on the cancel button to get out of edit mode. #108 has been added to the backlog and initially scheduled for Sprint 5.

Related work items: #103