Merged PR 20: Add roles and reorganize identity/user mgmt to new BC.
This includes a fair amount of experimentation, and includes reorganizing most identity and user management to a new bounded context, renaming that BC from Identity to UserMgmt, adding custom identity roles and an API and UI to display that list of roles. Finally, OpenApi and Swagger UI is added for initial experimentation. The API client is in a poor state, but is being left that way because it will be replaced in the next sprint with an auto-generated API client.