Merged PR 22: Add NSwag generated API client
This pull adds an API client code generator that builds API client and contract classes into the new project/assembly SSar.Presentation.ApiClient.CSharp. It runs automatically on each build unless AutoGenerate is set to false in the client .csproj file on the dotnet build command line.

To disable client auto-generation on each build in the CI/CD pipeline it is recommended to either set an environment condition in the project file or modify the build command in the pipeline.