Merged PR 27: Add EntryControl bounded context with EntryControlUser
Adds a new bounded context, EntryControl, and the first aggregate root, EntryControlUser. It includes a builder for the aggregate and appropriate unit tests. This also introduces to new value objects, PhoneNumber and EmailAddress. Note that it does NOT use the PersonName. It appears the PersonName value object may be closely linked to the MemberManagement domain and maybe should be relocated there. Time will tell.