Merged PR 28: Add AggregateRoot concept and convert IDs from int to Guid.
Converts IDs for AppUser and AppRole from in to Guid, and and switched the IEntity ID from int to Guid. It then adds EntityBase and AggregateBase classes and implements them on the recent EntryControlUser aggregate.

Note that the MemberMgmt BC has not been updated. It was an original test vertical slice and which has been abandoned and will be deleted soon then recreated in the future.